+++ joes_giant_toolbox/all/string_cleaner.py
@@ -17,18 +17,18 @@ class StringCleaner:
     operations: dict
         A dictionary containing each individual string cleaning function
     verbose: str, optional (default: True)
-        A global parameter dictating the verbosity of every string cleaning function
+        A global parameter dictating the verbosity of the output
 
     Methods
     -------
     apply_sequential_operations
-        Sequentially applies 1 or more string cleaning operations to a given string
+        Sequentially applies multiple string cleaning operations to the same string
 
     Notes
     -----
     You can get help on a specific string cleaning function using the built-in python help() function:
     >>> string_cleaner = StringCleaner()
-    >>> help( string_cleaner.operations["to_lowercase"] )
+    >>> help( string_cleaner.operations["extract_domain_from_url"] )
 
     A single string cleaning operation can be accessed directly:
     >>> string_cleaner.operations["remove_punctuation"]("j!o@e#i$s%t^h&e&b*e(s)t")
@@ -36,39 +36,6 @@ class StringCleaner:
         ...or called via a chain consisting of only that 1 operation:
     >>> string_cleaner.apply_sequential_operations("j!o@e#i$s%t^h&e&b*e(s)t", operation_names_list=["remove_punctuation"])
     joeisthebest
-
-    Available String Cleaning Operations
-    ------------------------------------
-    extract_domain_from_url
-        Extracts the base domain from a given website URL
-    remove_words_or_phrases
-        Removes specific 'words' or 'phrases' (i.e. specific characters in a specific order) from a given string, possibly observing word boundaries
-    to_lowercase
-        Converts all upper case characters in the given string to lower case
-    punctuation_to_spaces
-        Turns any character which is not a letter or a number into a space character
-    remove_punctuation
-        Removes any character which is not a letter or a number
-    numbers_to_spaces
-        Turns all number characters in a given string into space characters
-    remove_numbers
-        Removes all number characters from a given string
-    remove_spaces
-        Removes all spaces from a given string
-    newlines_to_spaces
-        Turn newline characters ("\\n") into space characters in a given string
-    multiple_spaces_to_single_spaces
-        Turns every sequence of space characters in a given string into a single space character
-    remove_spaces_at_start_and_end
-        Remove any spaces characters
-    non_letters_to_spaces
-        Turn every character in a given string which is not a letter into a single space character
-    remove_non_letters
-        Remove every character in a given string which is not a letter
-    join_single_space_separated_letters_together
-        Removes the space characters between sequences of single letter characters separated by a single space (e.g. "a  B c D  e" becomes "a  BcD  e")
-    remove_non_printable_ascii_chars
-        Removes all characters from a given string which are not in the python constant 'string.printable'
     """
 
     def __init__(self, verbose: bool = True) -> None:
@@ -91,7 +58,7 @@ class StringCleaner:
 # extract domain from URL #
 " ".join(
     re.findall(
-        "(?P<url>https?://[^/]+)",
+        "(?P<url>https?://[^/?]+)",
         "{raw_str}",
     )     
 )           
@@ -99,7 +66,7 @@ class StringCleaner:
                 )
             return " ".join(
                 re.findall(
-                    "(?P<url>https?://[^/]+)",
+                    "(?P<url>https?://[^/?]+)",
                     raw_str,
                 )
             )
@@ -181,7 +148,7 @@ re.sub(
             ' Joe is the best '
             """
             if verbose:
-                print(f're.sub(r"[^A-Za-z0-9 ]+", " ", "{raw_str}")')
+                print(f're.sub(r"[^A-Za-z0-9]+", " ", "{raw_str}")')
             return re.sub(r"[^A-Za-z0-9 ]+", " ", raw_str)
 
         def remove_punctuation(raw_str):
@@ -190,11 +157,11 @@ re.sub(
             Example Usage
             -------------
             >>> string_cleaner = StringCleaner()
-            >>> string_cleaner.operations["remove_punctuation"]("!@Joe#$Is%^The_Best&*")
-            'JoeIsTheBest'
+            >>> string_cleaner.operations["remove_punctuation"]("!@Joe# $is% ^the_ :best&*")
+            'Joe is the best'
             """
             if verbose:
-                print(f're.sub(r"[^A-Za-z0-9 ]+", "", "{raw_str}")')
+                print(f're.sub(r"[^A-Za-z0-9]+", "", "{raw_str}")')
             return re.sub(r"[^A-Za-z0-9 ]+", "", raw_str)
 
         def numbers_to_spaces(raw_str):
@@ -202,8 +169,14 @@ re.sub(
 
             Example Usage
             -------------
-            TODO
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["numbers_to_spaces"]("time2go2number136street")
+            'time go number   street'
             """
+            if verbose:
+                print(
+                    f'"{raw_str}".translate(str.maketrans("0123456789", " " * len("0123456789")))'
+                )
             return raw_str.translate(
                 str.maketrans("0123456789", " " * len("0123456789"))
             )
@@ -213,58 +186,142 @@ re.sub(
 
             Example Usage
             -------------
-            TODO
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["remove_numbers"]("time2go2number136street")
+            'timegonumberstreet'
             """
+            if verbose:
+                print(f'"{raw_str}".translate(str.maketrans("", "", "0123456789"))')
             return raw_str.translate(str.maketrans("", "", "0123456789"))
 
         def remove_spaces(raw_str):
+            """Removes all spaces from a given string
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["remove_spaces"]("U  S  A")
+            'USA'
+            """
+            if verbose:
+                print(f're.sub(" ", "", "{raw_str}")')
             return re.sub(" ", "", raw_str)
 
         def newlines_to_spaces(raw_str):
+            """Turn newline characters ("\\n") into space characters in a given string
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["newlines_to_spaces"]("\\nU\\nS\\nS\\nR\\n")
+            ' U S S R '
+            """
+            if verbose:
+                print(f'"{raw_str}".replace("\\n", " ")')
             return raw_str.replace("\n", " ")
 
         def multiple_spaces_to_single_spaces(raw_str):
+            """Turns every sequence of space characters in a given string into a single space character
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["multiple_spaces_to_single_spaces"]("the quick  brown   fox     jumped        over")
+            'the quick brown fox jumped over'
+            """
+            if verbose:
+                print(f're.sub(" +", " ", "{raw_str}")')
             return re.sub(" +", " ", raw_str)
 
-        def remove_spaces_at_start_and_end(raw_str):
+        def remove_whitespace_at_start_and_end(raw_str):
+            """Remove all space and newline characters at the start and end of a string
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["remove_whitespace_at_start_and_end"]("\\n\\n text of interest   ")
+            'text of interest'
+            """
+            if verbose:
+                print(f'"{raw_str}".strip()')
             return raw_str.strip()
 
         def non_letters_to_spaces(raw_str):
+            """Turn all non-letter (i.e. not A-Z) characters into space characters in given string
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["non_letters_to_spaces"]("  joe123is-:the8==|>best  ")
+            '  joe   is  the     best  '
+            """
+            if verbose:
+                print(f're.sub("[^a-zA-Z]", " ", "{raw_str}")')
             return re.sub("[^a-zA-Z]", " ", raw_str)
 
         def remove_non_letters(raw_str):
+            """Remove all non-letter (i.e. not A-Z) characters in given string
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["remove_non_letters"]("  joe123is-:the8==|>best  ")
+            'joeisthebest'
+            """
+            if verbose:
+                print(f're.sub("[^a-zA-Z]", "", "{raw_str}")')
             return re.sub("[^a-zA-Z]", "", raw_str)
 
         def join_single_space_separated_letters_together(raw_str):
+            """Removes the space characters between sequences of single letter characters separated by a single space
+            (e.g. "a  B c D  e" becomes "a  BcD  e")
+
+            Example Usage
+            -------------
+            >>> string_cleaner = StringCleaner()
+            >>> string_cleaner.operations["join_single_space_separated_letters_together"]("a  B c D ef  g h  ")
+            'a  BcD ef  gh  '
+            """
+            if verbose:
+                print(f're.sub(r"\\b([a-zA-Z]) (?=[a-zA-Z]\\b)", r"\\1", "{raw_str}")')
             return re.sub(r"\b([a-zA-Z]) (?=[a-zA-Z]\b)", r"\1", raw_str)
 
         def remove_non_printable_ascii_chars(raw_str):
-            """Removes all characters from a given string which are not in the python constant 'string.printable'"""
+            """Removes all characters from a given string which are not in the python constant 'string.printable'
+
+            Example Usage
+            -------------
+            TODO
+            """
+            if verbose:
+                print(
+                    f'"".join(filter(lambda x: x in self.printable_ascii_chars, {raw_str}))'
+                )
             return "".join(filter(lambda x: x in self.printable_ascii_chars, raw_str))
 
         self.operations["extract_domain_from_url"] = extract_domain_from_url
-        self.operations["remove_words_or_phrases"] = remove_words_or_phrases
-        self.operations["to_lowercase"] = to_lowercase
-        self.operations["punctuation_to_spaces"] = punctuation_to_spaces
-        self.operations["remove_punctuation"] = remove_punctuation
-        self.operations["numbers_to_spaces"] = numbers_to_spaces
-        self.operations["remove_numbers"] = remove_numbers
-        self.operations["remove_spaces"] = remove_spaces
-        self.operations["newlines_to_spaces"] = newlines_to_spaces
-        self.operations[
-            "multiple_spaces_to_single_spaces"
-        ] = multiple_spaces_to_single_spaces
-        self.operations[
-            "remove_spaces_at_start_and_end"
-        ] = remove_spaces_at_start_and_end
-        self.operations["non_letters_to_spaces"] = non_letters_to_spaces
-        self.operations["remove_non_letters"] = remove_non_letters
         self.operations[
             "join_single_space_separated_letters_together"
         ] = join_single_space_separated_letters_together
         self.operations[
+            "multiple_spaces_to_single_spaces"
+        ] = multiple_spaces_to_single_spaces
+        self.operations["newlines_to_spaces"] = newlines_to_spaces
+        self.operations["non_letters_to_spaces"] = non_letters_to_spaces
+        self.operations["numbers_to_spaces"] = numbers_to_spaces
+        self.operations["punctuation_to_spaces"] = punctuation_to_spaces
+        self.operations["remove_non_letters"] = remove_non_letters
+        self.operations[
             "remove_non_printable_ascii_chars"
         ] = remove_non_printable_ascii_chars
+        self.operations["remove_numbers"] = remove_numbers
+        self.operations["remove_punctuation"] = remove_punctuation
+        self.operations["remove_spaces"] = remove_spaces
+        self.operations[
+            "remove_whitespace_at_start_and_end"
+        ] = remove_whitespace_at_start_and_end
+        self.operations["remove_words_or_phrases"] = remove_words_or_phrases
+        self.operations["to_lowercase"] = to_lowercase
 
         if self.verbose:
             print("-- Initiated StringCleaner() --")