+++ joes_giant_toolbox/text.py
@@ -1,3 +1,5 @@
+"""Tools for working with text data"""
+
 from joes_giant_toolbox.all.longest_sentence_subsequence_plagiarism_detector import (
     longest_sentence_subsequence_plagiarism_detector,
 )