+++ test.py
@@ -1,3 +1,0 @@
-import joes_giant_toolbox.google_cloud
-
-help(joes_giant_toolbox.google_cloud)