+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 * Added documentation to **joes_giant_toolbox.google_cloud**
 
+* Added documentation to **joes_giant_toolbox.text**
+
 * Added package dependency group [google]
+
+* Improved documentation of **joes_giant_toolbox.text.StringCleaner**
 
 * Improved (verbose=True) output of function **joes_giant_toolbox.google_cloud.move_or_rename_file_in_gcloud_bucket**